# Embed.__len__ optimizations