## What type of pull request is this?

<!-- Check whichever applies to your PR -->
- [x] Non-breaking code change
- [ ] Breaking code change
- [ ] Documentation change/addition

## Description
Taking into account that embeds are constantly using the `__len__` property for validators whenever they are modified, the implementation of maps and optimal methods can help to make the validation faster which benefits with a gain 20-25% in comparison to the usual approach.

### Benchmark (tested with timeit, `n=1000000`)

| Embed | Default `__len__` | Modified `__len__` |
| - | - | - |
| Empty Embed | 0.2745 s | 0.1712 s |
| All parameters | 4.8217 s | 4.2479 s |
| All parameters (No Fields) | 0.7127 s | 0.6124 s

## Changes
<!-- - A bullet pointed list outlining the changes you have made -->
• Modified `Embed.__len__`

## Checklist

<!-- These are actions you **must** have taken, if you haven't, your PR will be rejected -->
- [x] I've formatted my code with [Black](https://black.readthedocs.io/en/stable/)
- [x] I've ensured my code works on `Python 3.10.x`
- [x] I've tested my code