@artem30801 2022-01-14
Looks nice
Does the walrus operator boost performance? Have you measured it in comparison to bare ifs and attribute re-usage?
Also, you should probably delete that in-code comment left for the older implementation X)