@Bluenix2 2022-01-14
After generating the length, why not save it to an attribute and whenever the embed is modified set it to None?

It'll make the `__len__` lookup instant if nothing changed.