@LordOfPolls 2022-01-14
@artem30801 
> you should probably delete that in-code comment left for the older implementation X)
arguably i think it still applies 😜