@LordOfPolls 2022-01-14
@Bluenix2 
> After generating the length, why not save it to an attribute and whenever the embed is modified set it to None?

It was my understanding *most* python objects cache their len