@Vioshim 2022-01-14
> Looks nice Does the walrus operator boost performance? Have you measured it in comparison to bare ifs and attribute re-usage? Also, you should probably delete that in-code comment left for the older implementation X)

I've been making multiple tests and it always gave better results to use the walrus operator instead of the if, I suppose it's by the fact that the If has to first execute the "function" or value of both sides and then assign it 

aka `a if condition else b` while using the walrus allows to only use the len function if it's needed, otherwise the function is never called which allows that small boost in speed.