@Vioshim 2022-01-14
> Would that perform better? Worse? Same? Could you please test it?

@artem30801 I've tested and seems to perform actually faster, most likely due to not having to use extra variables.

I'll proceed to update the PR