@Bluenix2 2022-01-14
> > After generating the length, why not save it to an attribute and whenever the embed is modified set it to None?
> 
> It was my understanding _most_ python objects cache their len

Yes, they do (the strings, lists and other built-in structures) but the whole body of `__len__` for Embed will still execute:
```python
>>> class AlwaysZero:
...     def __len__(self) -> int:
...             print('Accessed')
...             return 0
...
>>> zero = AlwaysZero()
>>> for _ in range(10):
...     len(zero)
...
Accessed
0
Accessed
0
Accessed
0
Accessed
0
Accessed
0
Accessed
0
Accessed
0
Accessed
0
Accessed
0
Accessed
0
```

It is up to us to cache the result.