@LordOfPolls 2022-01-14
> Yes, they do (the strings, lists and other built-in structures) but the whole body of `__len__` for Embed will still execute:

Oh i see what you mean, i thought you meant caching each len, my bad for not reading correctly